+++ lib/modules/console/console_assertion.ts
@@ -1,7 +1,7 @@
-import * as assertUtils from '../../utils/assert_utils';
 import BrowserConsole from './browser_console';
 import { ConsoleFilter } from './types';
 import { isNumber } from '../../utils/utils';
+import { AssertionError } from '../../errors';
 
 function processMessage(filterOptions: ConsoleFilter, count: number, actualCount: number): string {
     const filterMessage = processFilterMessage(filterOptions);
@@ -19,18 +19,17 @@ function processFilterMessage(filterOptions: ConsoleFilter): string {
     return `${typeMsg}${textMsg}`;
 }
 
-export default function(consoleModule: BrowserConsole, filterOptions: ConsoleFilter, count?: number, msg?: string): Promise<void> {
+export default async function(consoleModule: BrowserConsole, filterOptions: ConsoleFilter, count?: number, msg?: string): Promise<void> {
     const logs = consoleModule.filter(filterOptions);
     if (!isNumber(count)) {
         if (logs.length <= 0) {
             if (!msg) msg = processMessageWithoutExpectedCount(filterOptions);
-            return assertUtils.rejectAssertion("assert.console", msg);
+            throw new AssertionError("assert.console", msg);
         }
     } else {
         if (logs.length !== count) {
             if (!msg) msg = processMessage(filterOptions, count, logs.length);
-            return assertUtils.rejectAssertion("assert.console", msg);
+            throw new AssertionError("assert.console", msg);
         }
     }
-    return Promise.resolve();
 }